Adição do GlobalExceptionHandler à API e tratamento de exceções. OBS: existem pontos a melhorar, como o tratamento das exceções da DadosAusentesException.

+++ src/main/java/com/example/livros/service/LivroService.java
@@ -1,30 +1,50 @@
 package com.example.livros.service;
 
+import com.example.livros.exception.DadosAusentesException;
+import com.example.livros.exception.LivroNaoEncontradoException;
 import com.example.livros.repository.LivroRepository;
 import com.example.livros.model.Livro;
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import com.example.livros.exception.LivroNaoEncontradoException;
 
 @Service
 public class LivroService {
     @Autowired
     private LivroRepository livroRepository;
 
+    private boolean findMissingData(Livro livro) {
+        return livro.getTitulo() == null || livro.getAutor() == null;
+    }
+
     public List<Livro> buscarLivros() {
         return livroRepository.findAll();
     }
 
-    public Livro buscarLivroPorId(Long id) {
-        return livroRepository.findById(id).orElseThrow(() -> new RuntimeException("Usuário não encontrado."));
+    public Livro buscarLivroPorId(Long id) throws LivroNaoEncontradoException {
+        return livroRepository.findById(id).
+                orElseThrow(() -> new LivroNaoEncontradoException("Livro com o ID " + id + " não econtrado"));
     }
 
-    public Livro salvarLivro(Livro livro) {
+    public Livro salvarLivro(Livro livro) throws DadosAusentesException {
+        if(this.findMissingData(livro)) {
+            throw new DadosAusentesException("Opa, algum campo não foi informado.");
+        }
         return livroRepository.save(livro);
     }
 
-    public Livro atualizarLivro(Long id, Livro livroAtualizado) {
-        Livro livro = livroRepository.findById(id).orElseThrow(() -> new RuntimeException("Usuário não encontrado"));
+    public Livro atualizarLivro(Long id, Livro livroAtualizado) throws LivroNaoEncontradoException, DadosAusentesException {
+        if(!livroRepository.existsById(id)) {
+            throw new LivroNaoEncontradoException("Livro com o ID " + id + " não encontrado");
+        }
+
+        Livro livro = livroRepository.findById(id).
+                orElseThrow(() -> new LivroNaoEncontradoException("Livro com o ID " + id + " não econtrado"));
+
+        if(this.findMissingData(livroAtualizado)) {
+            throw new DadosAusentesException("Opa, algum campo não foi informado.");
+        }
 
         livro.setTitulo(livroAtualizado.getTitulo());
         livro.setAutor(livroAtualizado.getAutor());
@@ -32,7 +52,10 @@ public class LivroService {
         return livroRepository.save(livro);
     }
 
-    public void deletarLivro(Long id) {
+    public void deletarLivro(Long id) throws LivroNaoEncontradoException {
+        if(!livroRepository.existsById(id)) {
+            throw new LivroNaoEncontradoException("Livro com o ID " + id + " não encontrado");
+        }
         livroRepository.deleteById(id);
     }
 }
